requirements.txt
# Python Requirements for Image Migration
firebase-admin>=6.0.0
pillow>=10.0.0
